TheOrgBook/tob-api/requirements.txt
# Main
django>=1.11.5,<2
djangorestframework>=3.6.4,<3.8
djangorestframework-bulk>=0.2.1,<1
django-cors-headers>=2.1.0,<3

# Database managment
psycopg2>=2.7.3.1,<3

# Web Server
gunicorn>=19.7.1,<20
whitenoise>=3.3.1,<4

# Documentation
django_rest_swagger>=2.1.2,<3

# Testing
pytest-cov>=2.5.1,<3
pytest-django>=3.1.2,<4

# Search
drf-haystack>=1.6.1,<2
django-haystack>=2.7.dev0,<3
pysolr>=3.6.0,<4

# Von Agent
# git+https://github.com/nrempel/von_agent.git#von-agent
python3-indy==1.3.1-dev-408
von-agent==0.5.0
asyncio>=3.4.3,<4

